+++ tools/chapterNavigation/tool.js
@@ -32,8 +32,9 @@
       const $=parent.$;
       const study=parent.lichess.analysis?.study;
       if (!study) return;
+      if (!$('div.study__side .study__chapters').length) return;
       const trans=parent.translator;
-      const list=study.chapters.list();
+      const list=study.chapters.list.all();
       let container=$('div.study__side div[role="footer"]');
       if (!container.length&&list.length>1) {
         container=$(`<div role="footer">
@@ -71,7 +72,7 @@
         parent.global.console.warn('Could not determine chapterId');
         return;
       }
-      const list=study.chapters.list();
+      const list=study.chapters.list.all();
       let index=list.findIndex(c=>c.id==chapterId);
       const act=$(ev.target).attr('data-act');
       switch(act) {