+++ tools/explorerGambits/gambits.js
@@ -2933,12 +2933,12 @@
       ]
     },
     "r1bqkbnrpppp1ppp2n54p32B1P35N2PPPP1PPPRNBQK2RbKQkq-": {
-      "total": 25,
+      "total": 26,
       "moves": [
         {
           "uci": "f7f5",
           "san": "f5",
-          "nr": 1
+          "nr": 2
         },
         {
           "uci": "f8c5",
@@ -3033,12 +3033,12 @@
       ]
     },
     "r1bqkbnrpppp1ppp2n54p34P35N2PPPP1PPPRNBQKB1RwKQkq-": {
-      "total": 59,
+      "total": 60,
       "moves": [
         {
           "uci": "f1c4",
           "san": "Bc4",
-          "nr": 25
+          "nr": 26
         },
         {
           "uci": "f1b5",
@@ -3308,8 +3308,13 @@
       ]
     },
     "r1bqkbnrpppp2pp2n54pp22B1P35N2PPPP1PPPRNBQK2RwKQkq-": {
-      "total": 1,
+      "total": 2,
       "moves": [
+        {
+          "uci": "d2d4",
+          "san": "d4",
+          "nr": 1
+        },
         {
           "uci": "e4f5",
           "san": "exf5",
@@ -13753,17 +13758,17 @@
       ]
     },
     "rnbqkbnrpppp1ppp84p34P35N2PPPP1PPPRNBQKB1RbKQkq-": {
-      "total": 81,
+      "total": 82,
       "moves": [
+        {
+          "uci": "b8c6",
+          "san": "Nc6",
+          "nr": 60
+        },
         {
           "uci": "d7d5",
           "san": "d5",
           "nr": 1
-        },
-        {
-          "uci": "b8c6",
-          "san": "Nc6",
-          "nr": 59
         },
         {
           "uci": "g8f6",
@@ -13793,12 +13798,12 @@
       ]
     },
     "rnbqkbnrpppp1ppp84p34P38PPPP1PPPRNBQKBNRwKQkq-": {
-      "total": 203,
+      "total": 204,
       "moves": [
         {
           "uci": "g1f3",
           "san": "Nf3",
-          "nr": 81
+          "nr": 82
         },
         {
           "uci": "f1e2",
@@ -15108,17 +15113,17 @@
       ]
     },
     "rnbqkbnrpppppppp884P38PPPP1PPPRNBQKBNRbKQkq-": {
-      "total": 324,
+      "total": 325,
       "moves": [
+        {
+          "uci": "e7e5",
+          "san": "e5",
+          "nr": 204
+        },
         {
           "uci": "e7e6",
           "san": "e6",
           "nr": 22
-        },
-        {
-          "uci": "e7e5",
-          "san": "e5",
-          "nr": 203
         },
         {
           "uci": "c7c5",
@@ -15368,12 +15373,12 @@
       ]
     },
     "rnbqkbnrpppppppp8888PPPPPPPPRNBQKBNRwKQkq-": {
-      "total": 519,
+      "total": 520,
       "moves": [
         {
           "uci": "e2e4",
           "san": "e4",
-          "nr": 324
+          "nr": 325
         },
         {
           "uci": "d2d4",