+++ tools/extraChart/tool.js
@@ -532,9 +532,9 @@
           if (this.inCheck(p2.fen)) return 0;
           if (this.isPromotion(p1)) return 0;
           const move={
-            sx: p1.uci.charCodeAt(0)-97, 
+            sx: 104-p1.uci.charCodeAt(0), 
             sy: p1.uci.charCodeAt(1)-49,
-            x: p1.uci.charCodeAt(2)-97, 
+            x: 104-p1.uci.charCodeAt(2), 
             y: p1.uci.charCodeAt(3)-49
           };
           let board=parent.getBoardFromFen(p2.fen);
@@ -663,6 +663,7 @@
                  const elems=this.getInterestingMoveElements(color);
                  if (!this.colors.beforeInterestingMoves) this.colors.beforeInterestingMoves=dataset.hoverBackgroundColor||this.colors.originalChart;
                  dataset.hoverBackgroundColor=this.colors.interestingMoves();
+                 dataset.pointHoverBackgroundColor=this.colors.interestingMoves();
                  chart.setActiveElements(elems);
                  chart.update('none');
                })
@@ -673,6 +674,7 @@
                  if (!dataset) return;
                  const elems=[];
                  dataset.hoverBackgroundColor=this.colors.beforeInterestingMoves;
+                 dataset.pointHoverBackgroundColor=this.colors.beforeInterestingMoves;
                  chart.setActiveElements(elems);
                  chart.update('none');
                })
@@ -713,6 +715,7 @@
         const elements=dataset.data.filter(d=>parent.random()<0.3).map(d=>{ return { datasetIndex:0, index:d.x-1 }; });
         const color=colors[Math.round(parent.random()*colors.length)];
         dataset.hoverBackgroundColor=color;
+        dataset.pointHoverBackgroundColor=color;
         chart.setActiveElements(elements);
         chart.update('none');
         xElem.css('color',color);
@@ -720,6 +723,7 @@
       }
       parent.global.setTimeout(()=>xElem.remove(),1000);
       dataset.hoverBackgroundColor=initHoverBackgroundColor;
+      dataset.pointHoverBackgroundColor=initHoverBackgroundColor;
       chart.setActiveElements(initActiveElements);
       chart.update('none');
     };