+++ tools/mayTenth/tool.js
@@ -33,12 +33,15 @@
       const value=parent.currentOptions.getValue('mayTenth');
       this.logOption('LT day', value);
       $.cached('body').removeClass('lichessTools-mayTenth');
-      $('h1.site-title a').removeAttr('title');
+      $('a.site-title,#topnav section a:has(span.home)').removeAttr('title');
       if (!value) return;
+      const isMayTenth=new Date().toISOString().includes('-05-10');
       $.cached('body')
-        .toggleClass('lichessTools-mayTenth',new Date().toISOString().includes('-05-10'));
-      $('h1.site-title a')
-       .attr('title',trans.noarg('mayTenthTitle'));
+        .toggleClass('lichessTools-mayTenth',isMayTenth);
+      if (isMayTenth) {
+        $('a.site-title,#topnav section a:has(span.home)')
+          .attr('title',trans.noarg('mayTenthTitle'));
+      }
     }
   }
   LiChessTools.Tools.MayTenth=MayTenthTool;